feat: Include rules from `eslint-plugin` in `eslint-config` (#135)
##
[AMA-279](https://linear.app/clipboardhealth/issue/AMA-279/write-an-eslint-rule-for-checking-that-new-endpoints-in-backend-main)

Changes
---
- This is a follow-up to #131 where a new `eslint-plugin` package was
introduced containing definitions for custom ESLint rules.
- This PR incorporates that plugin into the config exported by
`@clipboard-health/eslint-config` so that using the config results in
the rules in the plugin being automatically enabled. Consuming project
do not need to directly install the `eslint-plugin` package.
- The PR also skips running the linter on the linter config itself since
the linter errors are not very useful and cause unavoidable typing
issues to appear on a .js file.