revert: handle recipient chunking in NotificationClient (#211)
Summary
===
Reverts the chunking aspect of
https://github.com/ClipboardHealth/core-utils/pull/209.

Knock will throw if you use the same idempotency key, but the body
changes. However, we moved the body creation to the jobs themselves so
that if, for example, one of your message template variables changes,
you're getting the most up-to-date value for it.

This means that a job to a bunch of recipients could run. Say the first
batch of 1000 succeeds, and the 2nd fails because of a network issue. We
retry the job, but in the meantime, a variable changes in the first
batch. Now the job will fail indefinitely and the later batches won't
get notified.

We need each chunk to be in its own background job so it either succeeds
or fails and retries.

Co-authored-by: Rocky Warren <rocky.warren@clipboardhealth.com>