feat: make it harder to misuse NotificationsClient (#215)
Summary
===
People are having a hard time using `NotificationClient` properly, which
points to a failure in its APIs. While auditing `.trigger` calls, I'm
seeing the following problems:

Problems:
1. Clients must remember to hash the notification request body in the
background job if they pull template variables from a database so that
if they change, the idempotency key gets updated, preventing Knock's API
from throwing.
2. `idempotencyKey` is meant to prevent duplicate notifications. Clients
aren't creating good ones. I've seen adding `new Date()` in the key,
which means it's always different and doesn't prevent duplicates or the
opposite, where it's not unique enough and prevents customers from
getting notifications they otherwise should have. For example, clients
may only include the `recipientId` as the key, but recipients could
reasonably receive the notification more than once within Knock's 24
hour idempotency key timeout.
3. Clients must remember to call `enqueueOneOrMore` in our code instead
of the standard background job enqueue method or we won't properly chunk
recipients that are greater than Knock's limit so that each job is only
operating on a single chunk (which helps with reties, as detailed
[here](https://github.com/ClipboardHealth/core-utils/pull/211#issue-3481902719)).
4. `expiresAt` is meant to prevent stale notifications. Clients aren't
setting it properly. They set it in the job itself instead of at enqueue
time, which means it always in the future.
5. Our background job libraries also have idempotency keys, but clients
aren't reliably setting them.
6. `.trigger` accept `attempt`, but more often than not, clients are
hard-coding it to `1` instead of using the background job's attempt
number, harming debuggability.

This PR is an attempt to fix all of them:
1. We could hash the trigger body in `NotificationClient` so library
clients don't need to remember (or provide a flag like
includeTriggerBodyHash` in `IdempotencyKey`). But I don't think this is
required. Since we enforce going through `enqueueOneOrMore`, each job
should only contain one notification chunk that will either succeed or
fail. If it fails, the idempotency key isn't stored and if it succeeds
and we try again with a different body, we want Knock to throw an error.
2. Create an `IdempotencyKey` class with documentation on what to
include in the key. Include other details automatically so clients can't
forget.
3. Require an internal-only class in `.trigger` so clients must to go
through `enqueueOneOrMore` to create the proper trigger request body.
4. Required by `enqueueOneOrMore`.
5. Set in `enqueueOneOrMore`.
6. Though not foolproof, I created an example that shows proper usage we
can set up Sourcegraph monitoring to alert on `attempt: 1`.

In a follow-up PR, I'll make the following breaking changes:
1. Remove `TriggerRequest.key` and make `TriggerRequest.workflowKey`
required.
2. Remove the `string` option for `TriggerRequest.idempotencyKey`.

---------

Co-authored-by: Rocky Warren <rocky.warren@clipboardhealth.com>