fix: use branded type instead of class for serialzation and simplicity (#220)
Summary
===
Serializing/deserializing the `TriggerIdempotencyKey` class meant it
lost its `instance` and became a plain object, which meant I had to add
`toIdempotencyKey` in `notificationClient` to rebuild it. This could
cause issues if others try to use it as a class prior to calling
`.trigger` and potentially sets a bad precedence of passing classes to
background jobs. This PR converts it to a plain object, while
maintaining the goal of having people call `.trigger` by way of
`.enqueueOneOrMore`.

---------

Co-authored-by: Rocky Warren <rocky.warren@clipboardhealth.com>