feat: Introduce mongo-jobs (#224)
Moving our mongo based background jobs to core-utils. 
I went with `@clipboard-health/mongo-jobs` name as it's much easier to
say/type than our old `background-jobs-mongo`.

In order to be able to run tests for the project I had to introduce
mongo running in docker in CI.

The functionality is copied from our internal package. 
I made some improvements to code nudged by coderabbit, but the
underlying logic is the same.