feat: Rename BackgroundJobService to BackgroundJobs (#242)
Initially I wanted to call it `MongoJobs` but then @kepten convinced me to change it into `BackgroundJobs`. The `...Service` part doesn't have much meaning and it's just making the name longer and more annoying to write and read. And `import {BackgroundJobs } from @clipboard-health/mongo-jobs` has a nice ring to it.