fix: prevent duplicate key errors from aborting transactions (#274)
Summary
===

When we enqueue a job we detect duplicate key errors, swallow the error,
and return `undefined`; but since MongoDB aborts transactions when there
is a `DuplicateKeyError` this causes our current transaction to be
aborted.

To fix this we should check if there is a duplicate before inserting,
but only if we are inserting a job within a transaction and with a
`uniqueKey`

Changes
---
- Add a test which fails because of the transaction being aborted
- Make the test pass by adding a check for existing jobs with the
`uniqueKey`.

Notes
---

I initially thought this could be achieved by setting `ordered: false`
as an option on the `create` and avoid any additional queries, but that
only prevents `DuplicateKeyError` from aborting the transaction if we
insert the job with `writeConvern = 0` which we don't want to do.

This solution requires an additional query when enqueuing, but only when
enqueing with a `uniqueKey` and inside a transaction, and it is a quick
lookup using a unique index,