refactor: Forbid extra boolean casting (#286)
Summary
===

Enforces [Use isDefined() or, if needed Boolean(), not
!!](https://www.notion.so/BP-TypeScript-Style-Guide-5d4c24aea08a4b9f9feb03550f2c5310?source=copy_link#2558643321f480398b5aed5c974fef3d)
via `"no-extra-boolean-cast"`


[Ticket](https://linear.app/clipboardhealth/issue/GRO-1075/no-extra-boolean-cast)

Ahead of adding this linting rule, I've already refactored backend main
to move from 90 to -> less than 30 instances of `!!` The ones that
remain are the slightly trickier cases, using !! on strings. It's tricky
because the current implementation returns false for !!"" and I didn't
want to modify those because it meant:
1. Turning those all to `isDefined` which has different behavior that
could be wrong
2. Codify all of those as `isDefined && length > 0`, in case that wasn't
necessary
3. Doing the deep research to understand if even empty strings are
possible cases for those usages.

Because we are on eslint v8, we can't use the stronger
`enforceForInnerExpressions: true` and instead have to rely on
`enforceForLogicalOperands: true`

When we upgrade to eslint v9, `enforceForLogicalOperands` becomes
deprecated and we should migrate over. The difference is that
`enforceForInnerExpressions` is slightly stronger and checks more
places.